feat: add support for adding invites with the new UI
