+++ shared/tenants/src/types.ts
@@ -20,7 +20,6 @@ export type TenantList = {
 
 export type InviteeDetails = {
   email: string;
-  role: string;
   code: string;
 };
